+++ sdk/communication/Azure.Communication.Common/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release History
 
 ## 1.0.0-beta.4 (Unreleased)
-
+### Breaking Changes
+- Renamed `CommunicationUserCredential` to `CommunicationTokenCredential`.
 
 ## 1.0.0-beta.3 (2020-11-16)
 Updated `Azure.Communication.Common` version.